solution/requirements.txt
# Core API
fastapi==0.128.0
uvicorn==0.28.0
pydantic==2.7.0
psycopg2-binary==2.9.9
confluent-kafka==2.3.0

# Testing
pytest==7.4.3
httpx==0.26.0
pytest-asyncio==0.23.0

# Utilities
python-dotenv==1.0.0
